requirements.txt
pyyaml>=6.0
kafka-python>=2.0.2
boto3>=1.24.0
sqlalchemy>=1.4.0
pandas>=1.4.0
python-dateutil>=2.8.2
tweepy>=4.10.0
jsonschema>=4.16.0
